Add files via upload
This proposal presents the development plan for Smart Resume Analyzer. This project uses Artificial Intelligence, specifically Natural Language Processing (NLP) and Machine Learning (ML), to automate resume screening, skill analysis, and job matching. It explains the core problem we're addressing, our research goals, and the methods we’ll be using, and it also looks into ethical and industry-related aspects. Overall, it lays the groundwork for how the system will be designed and built as part of this academic project.